Update basic_content.dart
When we use CircularProgressIndicator() directly, it takes up the entire card/container in which we are using CachedNetworkImage(). Hence, it doesn't look good to the user as it bloats over the container. So, it becomes better to wrap CircularProgressIndicator() inside Center() widget so it is just shown in the center of  the card/container.